forgot to set some sizes to uint16_t ...

+++ EasyTransfer/src/EasyTransfer.h
@@ -51,12 +51,12 @@ public:
 private:
     Stream *_stream;
     // NewSoftSerial *_serial;
-    uint8_t *address;     // address of struct
-    uint16_t size;        // size of struct
-    uint8_t *rx_buffer;   // address for temporary storage and parsing buffer
-    uint8_t rx_array_inx; // index for RX parsing buffer
-    uint8_t rx_len;       // RX packet length according to the packet
-    uint8_t calc_CS;      // calculated Chacksum
+    uint8_t *address;      // address of struct
+    uint16_t size;         // size of struct
+    uint8_t *rx_buffer;    // address for temporary storage and parsing buffer
+    uint16_t rx_array_inx; // index for RX parsing buffer
+    uint16_t rx_len;       // RX packet length according to the packet
+    uint8_t calc_CS;       // calculated Chacksum
 };
 
 #endif
